[MOB-260] moved minimal ad to a more appropriate package
