+++ app/src/main/java/cm/aptoide/pt/ads/AdsRepository.java
@@ -10,7 +10,6 @@ import android.content.SharedPreferences;
 import android.content.res.Resources;
 import android.net.ConnectivityManager;
 import cm.aptoide.accountmanager.AptoideAccountManager;
-import cm.aptoide.pt.database.room.MinimalAd;
 import cm.aptoide.pt.dataprovider.model.v2.GetAdsResponse;
 import cm.aptoide.pt.dataprovider.ws.v2.aptwords.AdsApplicationVersionCodeProvider;
 import cm.aptoide.pt.dataprovider.ws.v2.aptwords.GetAdsRequest;
@@ -199,8 +198,7 @@ public class AdsRepository {
                     .observe())));
   }
 
-  private Observable<MinimalAd> mapRandomAd(
-      Observable<GetAdsResponse> getAdsResponseObservable) {
+  private Observable<MinimalAd> mapRandomAd(Observable<GetAdsResponse> getAdsResponseObservable) {
     Random rand = new Random();
     return getAdsResponseObservable.map((getAdsResponse) -> getAdsResponse.getAds())
         .flatMap(ads -> {
