+++ app/src/main/java/cm/aptoide/pt/search/model/SearchAdResult.java
@@ -1,8 +1,8 @@
 package cm.aptoide.pt.search.model;
 
+import cm.aptoide.pt.ads.MinimalAd;
 import cm.aptoide.pt.ads.data.AptoideNativeAd;
 import cm.aptoide.pt.ads.data.Payout;
-import cm.aptoide.pt.database.room.MinimalAd;
 import org.parceler.Parcel;
 
 @Parcel public class SearchAdResult {
