+++ app/src/main/java/cm/aptoide/pt/view/recycler/widget/GridAdWidget.java
@@ -6,8 +6,8 @@ import android.widget.TextView;
 import androidx.fragment.app.FragmentActivity;
 import cm.aptoide.pt.AptoideApplication;
 import cm.aptoide.pt.R;
+import cm.aptoide.pt.ads.MinimalAd;
 import cm.aptoide.pt.crashreports.CrashReport;
-import cm.aptoide.pt.database.room.MinimalAd;
 import cm.aptoide.pt.networking.image.ImageLoader;
 import cm.aptoide.pt.search.model.SearchAdResult;
 import cm.aptoide.pt.view.recycler.displayable.GridAdDisplayable;
